+++ bdc/bdc/__init__.py
@@ -15,7 +15,7 @@ from gendbc import gendbc
 from db_edu_util.notebooktools import parse_source_notebook, NotebookError
 from db_edu_util import (databricks, wrap2stdout, error, verbose, set_verbosity,
                          warn, verbosity_is_enabled, info, die,
-                         EnhancedTextWrapper)
+                         EnhancedTextWrapper, working_directory)
 from db_edu_util.databricks import DatabricksError
 from grizzled.file import eglob
 from bdc.bdcutil import *
@@ -31,9 +31,9 @@ from typing import (Sequence, Any, Type, TypeVar, Set, Optional, Dict,
                     AnyStr, Tuple, NoReturn, Generator, Union, Callable, Set)
 
 __all__ = ['bdc_check_build', 'bdc_list_notebooks', 'bdc_build_course',
-           'bdc_download', 'bdc_upload', 'bdc_check_build',
+           'bdc_download', 'bdc_upload', 'bdc_check_build', 'bdc_load_build',
            'bdc_print_info', 'BuildError', 'UploadDownloadError',
-           'BuildConfigError', 'UnknownFieldsError']
+           'BuildConfigError', 'UnknownFieldsError', 'BuildData']
 
 # ---------------------------------------------------------------------------
 # Constants
@@ -41,7 +41,7 @@ __all__ = ['bdc_check_build', 'bdc_list_notebooks', 'bdc_build_course',
 # (Some constants are below the class definitions.)
 # ---------------------------------------------------------------------------
 
-VERSION = "1.36.0"
+VERSION = "1.37.0"
 
 DEFAULT_BUILD_FILE = 'build.yaml'
 PROG = os.path.basename(sys.argv[0])
@@ -839,6 +839,15 @@ class BuildData(DefaultStrMixin):
         ).substitute(
             folder_vars
         )
+
+    @property
+    def has_profiles(self):
+        """
+        Quick way to check whether the build has profiles or not.
+
+        :return: True or False
+        """
+        return len(self.profiles) > 0
 
     @property
     def notebooks(self):
@@ -3067,6 +3076,20 @@ def bdc_build_course(build_file: str,
         dest_dir = joinpath(os.getenv("HOME"), "tmp", "curriculum",
                             build.course_id)
     build_course(build, dest_dir, overwrite)
+
+
+def bdc_load_build(build_file: str) -> BuildData:
+    """
+    Load the build file and return the BuildData object. Useful for tools
+    outside bdc (e.g., course) that need information from the build.
+
+    :param build_file: path to the build file
+
+    :return: The parsed build data
+
+    :raises BuildConfigError: something is wrong with the build file
+    """
+    return load_and_validate(build_file)
 
 
 # ---------------------------------------------------------------------------